refactor(harness)!: consolidate the LangGraph harness, remove tracing handler
Collapse _langgraph_async / _langgraph_messages / _langgraph_tracing into
_langgraph_sync (emit_langgraph_messages, convert helpers) and _langgraph_turn
(stream_langgraph_events). Span tracing is now derived from the canonical
stream by UnifiedEmitter, so create_langgraph_tracing_handler is removed.

Public facade names are unchanged except the removed handler. All in-repo
consumers were migrated to the unified surface in the preceding PRs.

BREAKING CHANGE: create_langgraph_tracing_handler is removed.
